commit 28-01

+++ src/main/java/com/springboot/app/models/dao/IReservaDao.java
@@ -9,6 +9,6 @@ import com.springboot.app.models.entity.Reserva;
 * @author Jose
 */
 
-public interface IReservaDao extends JpaRepository<Reserva, Long> {
+public interface IReservaDao extends JpaRepository<Reserva, Integer> {
 
 }
